Update maintainer, description and add VCS info
The maintainer's name and email were different from the other Maratona
packages, so it had to be changed. Also, information about the project
main repository on Github has been added to the control file.

The package extended description was missing, which triggered Lintian's
extended-description-is-empty error. This has been fixed appending a
new description (progress on issue #2)

- updates maintainer's name and email
- adds VCS details
- adds extended description (fixes error extended-description-is-empty,
  see issue #2 for more details)